requirements.txt
wheel
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
marshmallow==3.20.1
python-abc
python-dotenv==1.0.0
SQLAlchemy==2.0.20
psycopg2-binary
pytz==2023.3
distlib==0.3.1
Faker==19.3.1
gunicorn
requests==2.31.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock
pytest-env